requirements.txt
Django==5.1.7
scikit-learn==1.6.1
mysqlclient==2.2.7
django-polymorphic==3.1.0
joblib==1.4.2
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==79.0.1
